Cambio del directorio
Todo en la carpeta principal
