+++ aplication/config/dbConfig.php
@@ -1,0 +1,13 @@
+<?php
+/**
+ * Created LuisAnc
+ * Date: 02/11/2017
+ * Variables de configuracion para a conexion con base de datos
+ */
+return array(
+    "host"      =>"localhost",
+    "user"      =>"root",
+    "pass"      =>"",
+    "database"  =>"softeedPr"
+);
+?>
