+++ aplication/config/mainConfig.php
@@ -1,0 +1,12 @@
+<?php
+/**
+ * Created by Social Nodes.
+ * User: LuisAnc
+ * Date: 02/11/2017
+ * Variables globales
+ */
+define("DEFAULT_CONTROLLER",'usuarios');
+define("DEFAULT_ACTION",'index');
+define("DEFAULT_LANG",'es');
+define("DEFAULT_PATH",__DIR__);
+?>
