+++ aplication/core/Info
@@ -1,0 +1,2 @@
+Vc:1.0.0
+En esta carpeta se encuentran las clases padre que contienen las funciones basicas
