+++ aplication/images/info
@@ -1,0 +1,3 @@
+Vc:1.0.0
+Esta carpeta esta destinada a guardar las imagenes necesarias para la aplicacion
+Nota:Unicamente las estaticas como logos e iconos
